pyproject.toml
[project]
name = "ice-colder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiomqtt>=2.3.0",
    "codeflash>=0.12.3",
    "fastapi[standard]>=0.115.11",
    "loguru>=0.7.3",
    "phonenumbers>=9.0.5",
    "pillow>=11.1.0",
    "pydantic>=2.10.6",
    "pydantic-extra-types>=2.10.4",
    "pydantic-settings>=2.9.1",
    "qrcode>=8.1",
    "ruff>=0.11.2",
    "transitions>=0.9.2",
]

[tool.codeflash]
# All paths are relative to this pyproject.toml's directory.
module-root = "."
tests-root = "tests"
test-framework = "pytest"
benchmarks-root = "tests\\benchmarks"
ignore-paths = []
disable-telemetry = false
formatter-cmds = ["ruff check --exit-zero --fix $file", "ruff format $file"]

[dependency-groups]
dev = [
    "pytest-asyncio>=1.3.0",
]
